backend/aws_lambda/requirements.txt
# AWS Lambda requirements
# Note: For CLIP processing, you'll need a Lambda container image with PyTorch

boto3==1.34.10
botocore==1.34.10
Pillow==10.1.0
numpy==1.24.3

# For Lambda container with CLIP:
# open-clip-torch==2.23.0
# torch==2.1.0 (CPU version)

